refactor(teal): Refactor atproto api implementation out of tight teal coupling
