+++ src/backend/server/auth.ts
@@ -195,16 +195,16 @@ export const setupAuthRoutes = (app: Express, logger: Logger, sourceMiddle: Expr
                 handle
             } = {}
         } = req;
-            const url = await (validClient.client as BlueSkyOauthApiClient).createAuthorizeUrl(handle as string);
+            const url = await (validClient.client.client as BlueSkyOauthApiClient).createAuthorizeUrl(handle as string);
             res.redirect(url)
         }
 
         if(intents[1].includes('client-metadata.json')) {
-            return res.json((validClient.client as BlueSkyOauthApiClient).getMetadata());
+            return res.json((validClient.client.client as BlueSkyOauthApiClient).getMetadata());
         }
 
         if(intents[1].includes('oauth/callback')) {
-            const result = await (validClient.client as BlueSkyOauthApiClient).handleCallback(new URLSearchParams(req.query as Record<string, string>));
+            const result = await (validClient.client.client as BlueSkyOauthApiClient).handleCallback(new URLSearchParams(req.query as Record<string, string>));
             if(result) {
                 return res.status(200);
             }
